Server, Signup, and Login
The server creates a database called "users" to store user signup data
in. The Signup code correctly stores a user in "users". The login file
correctly searched for that user and redirects to dashboard.html after
someone logs in. Problem: the login code isn't storing anything in local
storage as far as i can tell. If someone could look into this for me,
that would be great!