Sign up, Log in, about, contact, and Homepage all linked and working
All of the items are linked together and sign in and log in store the
correct information in the database and local storage. They also
redirect to /dashboard.html if a username and password already exists in
local storage.
Also improved the front-end for sign up and log in.

TODO on these pages: substitute a better logo image in for the current
logo (I made an editing mistake and left some white background in this
current picture), substitute the correct SwoleMate banner logo, generate
professional text for about and contact pages, make sure pages can scale
to other sized monitors and phones (framework?).